rename to assign and deletu unused stuff
