+++ src/stan/model/indexing.hpp
@@ -3,11 +3,9 @@
 
 #include <stan/model/indexing/deep_copy.hpp>
 #include <stan/model/indexing/index.hpp>
-#include <stan/model/indexing/index_list.hpp>
-#include <stan/model/indexing/lvalue.hpp>
-#include <stan/model/indexing/lvalue_varmat.hpp>
+#include <stan/model/indexing/assign.hpp>
+#include <stan/model/indexing/assign_varmat.hpp>
 #include <stan/model/indexing/rvalue.hpp>
 #include <stan/model/indexing/rvalue_varmat.hpp>
-#include <stan/model/indexing/rvalue_return.hpp>
 
 #endif
