+++ src/test/unit/model/indexing/assign_test.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <stdexcept>
 #include <vector>
-#include <stan/model/indexing/lvalue.hpp>
+#include <stan/model/indexing/assign.hpp>
 #include <stan/model/indexing/rvalue.hpp>
 #include <stan/math/rev.hpp>
 #include <gtest/gtest.h>
